Add ability to tangle to different files, and based on language
You can switch the file you're tangling to throughout the document. You
can specify the tangle file for a particular language by adding the
language as a parameter, i.e. :Tangle(python). The language parameter
has to match the language in the header of the block. If a file isn't
specified for a language, the general :Tangle file is used.